Remove redundant cast to byte

+++ src/InlineIL.Examples/Examples.cs
@@ -8,7 +8,7 @@ namespace InlineIL.Examples
             where T : struct
         {
             IL.Push(ref value);
-            IL.Push((byte)0);
+            IL.Push(0);
             IL.Emit(OpCodes.Sizeof, typeof(T));
             IL.Emit(OpCodes.Unaligned, 1);
             IL.Emit(OpCodes.Initblk);
